requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.5.1
astunparse==1.6.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
distro==1.9.0
exceptiongroup==1.2.2
flatbuffers==24.3.25
gast==0.4.0
geopandas>1.0.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
jiter==0.6.1
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
openai==1.51.2
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tqdm==4.66.5
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
